feat: handle random attack
