+++ src/constants/constants.ts
@@ -1,2 +1,3 @@
 export const HTTP_PORT = 8181;
 export const WS_PORT = 3000;
+export const FIELD_SIZE = 10;
